Move to webpack imports over require.ensure
This means the node code uses .babelrc and the webpack uses the babel config within the webpack.config files. This is required as I need webpack to process the imports for the front end code otherwise (if babel does it) code splitting does not work. But I still want babel to do it for the node code